Graduation
My high school graduation, a milestone I worked so hard for. As the firstborn, I’ve always felt the weight of expectation, but also the warmth of knowing I’m paving the way for the rest of the family. This photo isn’t just a celebration of my achievement, but a reflection of all the love, sacrifice, and encouragement you've given me.