New Orleans
This photo was taken during the beginning of our journey to New Mexico, right as we made a stop in New Orleans. We found reasons to laugh, bond, and make memories that still make us smile today. We all made our best attempt at copying Logan, and honestly, it turned into one of those goofy, unforgettable moments that captures exactly who we are as a family.