titanic
This photo was taken after a family hike, right at the top of the port. Mom and I couldn’t resist joking around and striking a pose like they do in *Titanic*, arms out, wind in our hair, totally dramatic and loving every second of it. It’s little moments like these that remind me how much fun we have together. No matter where we are, we always find a way to laugh, be silly, and make memories that stick.