the legend herself
I call this one “The Legend Herself.” This was when you were just starting out, fresh in the Air Force, a baby maintainer, and already showing the kind of strength most people spend a lifetime trying to find.

You had just stepped into a whole new world, balancing your role as a young mom and a service member. Looking back at this photo now, I see more than just a uniform or a serious face. I see power, purpose, and pride. You were doing what needed to be done for your family, and you did it with grace and grit, even when it was hard.

You were and still are a force to be reckoned with. This picture marks the beginning of a journey that shaped not just your future, but all of ours too. You’ve always been a legend in my eyes.